Update to Minecraft 1.21

+++ src/main/java/phanastrae/operation_starcleave/entity/projectile/SplashStarbleachEntity.java
@@ -1,21 +1,13 @@
 package phanastrae.operation_starcleave.entity.projectile;
 
-import net.minecraft.block.AbstractCandleBlock;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.CampfireBlock;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.FlyingItemEntity;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.projectile.thrown.ThrownItemEntity;
 import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
-import net.minecraft.potion.Potion;
-import net.minecraft.potion.PotionUtil;
-import net.minecraft.potion.Potions;
-import net.minecraft.registry.tag.BlockTags;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
@@ -26,14 +18,11 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.World;
-import net.minecraft.world.WorldEvents;
 import phanastrae.operation_starcleave.block.OperationStarcleaveBlocks;
 import phanastrae.operation_starcleave.entity.OperationStarcleaveEntityTypes;
 import phanastrae.operation_starcleave.item.OperationStarcleaveItems;
 import phanastrae.operation_starcleave.particle.OperationStarcleaveParticleTypes;
 import phanastrae.operation_starcleave.world.starbleach.Starbleach;
-
-import java.util.List;
 
 public class SplashStarbleachEntity extends ThrownItemEntity implements FlyingItemEntity {
 
@@ -88,7 +77,7 @@ public class SplashStarbleachEntity extends ThrownItemEntity implements FlyingIt
     }
 
     @Override
-    protected float getGravity() {
+    protected double getGravity() {
         return 0.05F;
     }
